# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: nethesis-insights
  version: "0.1.0"
  description: |
    Central log-anomaly analysis, Threat Shield and fleet-sizing server for a
    NethServer 8 fleet (~2700 nodes). Three independent pipelines are
    documented here, each its own binary behind one public host, with
    Traefik routing by path prefix and stripping the prefix before proxying:

    - **Log pipeline** (`insightsd`, prefix `/logs`: `/logs/v1/bundles`,
      `/logs/v1/findings`): edge nodes ship deduplicated, masked log bundles
      every 15 minutes; the server gates each one against novelty and
      deviation, calls an LLM only when the gate fires, and stores findings
      keyed by a server-computed fingerprint so the same problem is never
      re-raised.
    - **Threat Shield** (`threatd`, prefix `/blocklist`:
      `/blocklist/v1/events`, `/blocklist/v1/feed`,
      `/blocklist/v1/allowlist-requests`): a separate pipeline with no LLM
      and no fingerprint. Edges report CrowdSec ban decisions; the server
      publishes a fleet-consensus blocklist once an address is independently
      reported by enough distinct systems.
    - **Fleet sizing** (`sizingd`, prefix `/sizing`: `/sizing/v1/reports`): a
      third pipeline, also with no LLM, no gate, no fingerprint and no queue.
      NS8 cluster leaders post one complete-UTC-day workload and performance
      report per cluster; the server scores each node and publishes cohort
      hardware baselines.

...

...
servers:
...
paths:
...
  /logs/v1/findings:
    get:
      operationId: listFindings
      tags: [log-pipeline]
      summary: List findings for the authenticated system
      description: |
        Returns findings for the system named by the authenticated
        credential only -- there is no cross-system read on this surface.
        Findings are ordered severity-descending (`critical > high > medium
        > low`), then `last_seen` descending.
      security:
        - basicAuth: []
      parameters:
        - name: since
          in: query
          required: false
          description: Unix-millis lower bound on `last_seen`. Omit for no lower bound.
          schema: { type: integer, format: int64 }
        - name: status
          in: query
          required: false
          description: Filter to one status (`open` or `stale`). Omit for all statuses.
          schema: { type: string, enum: [open, stale] }
      responses:
        "200":
          description: The matching findings, possibly an empty list.
          content:
            application/json:
              schema:
                type: object
                required: [findings]
                properties:
                  findings:
                    type: array
                    items: { $ref: "#/components/schemas/Finding" }
              example:
                findings:
                  - id: 01J8Z1K2M3N4P5Q6R7S8T9V0W1
                    system_id: abc123
                    fingerprint: v1:9f2b7e1c4a...
                    severity: high
                    title: Repeated connection refused on traefik1
                    summary: traefik1 logged 37 connection-refused events in 15 minutes, far above baseline.
                    suggested_action: Check whether the upstream service on traefik1 is down or overloaded.
                    modules: [traefik1]
                    evidence:
                      - "<3> [n1:traefik1:traefik] connection refused to 10.0.0.4:8080"
                    status: open
                    occurrence_count: 1
                    first_seen: 1754380811000
                    last_seen: 1754381690000
                    llm_model: nvidia/nemotron-3-ultra-550b-a55b:free
                    prompt_version: v1
        "400":
          description: Invalid `since` (not an integer).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: "invalid since parameter" }
        "401":
          description: Invalid or missing credentials.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: "invalid credentials" }
        "405":
          description: Method other than `GET`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: "method not allowed" }
        "503":
          description: Forward-auth validator unreachable, or the store failed to read.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: "temporarily unavailable" }
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: >-
        `system_id:auth_token` HTTP Basic, forward-auth validated against
        `AUTH_VALIDATE_URL` (default `https://my.nethesis.it/auth`) with a
        TTL cache. Fail-closed: a rejected credential is `401`, a validator
        that cannot be reached is `503` (retryable), never a silent `401`.
    metricsAuth:
      type: http
      scheme: basic
      description: >-
        A single `prometheus` user, HTTP Basic, checked by Traefik's own
        `metrics-auth` middleware against `/etc/traefik/metrics.htpasswd` --
        never by the Go backends and never against `AUTH_VALIDATE_URL` or
        `ADMIN_API_KEY`. The password is generated at deploy time
        (`deploy/gen-metrics-auth.sh`) and lives in
        `/etc/insights/metrics.env`; see `docs/admin-guide.md`'s "Metrics"
        section.
  schemas:
    Error:
      type: object
      description: The uniform error body for every non-2xx JSON response on every surface.
      required: [error]
      properties:
        error: { type: string }
...
    Finding:
      type: object
      description: >-
        Server-computed and server-owned. Identity is the `fingerprint`
        (sha256 over sorted/deduped fields, never a model-authored string),
        so an inconsistently worded LLM restatement of the same problem
        collapses onto the same row instead of re-raising it.
      required:
        - id
        - system_id
        - fingerprint
        - severity
        - title
        - summary
        - suggested_action
        - modules
        - evidence
        - status
        - occurrence_count
        - first_seen
        - last_seen
        - llm_model
        - prompt_version
      properties:
        id: { type: string, description: ULID. }
        system_id: { type: string }
        fingerprint: { type: string, description: "sha256, `v1:`-prefixed." }
        severity: { type: string, enum: [critical, high, medium, low] }
        title: { type: string }
        summary: { type: string }
        suggested_action: { type: string }
        modules:
          type: array
          items: { type: string }
        evidence:
          type: array
          items: { type: string }
        status: { type: string, enum: [open, stale] }
        occurrence_count: { type: integer }
        first_seen: { type: integer, format: int64 }
        last_seen: { type: integer, format: int64 }
        reopened_at:
          type: ["integer", "null"]
          format: int64
          description: Set when a finding that had gone stale (absent past STALE_AFTER) recurred.
        llm_model: { type: string }
        prompt_version: { type: string, description: Code constant stamped at analysis time, never configuration. }
...